📁🔨 refactor(cart): CartSidebar 오타 수정, any 제거, getRemainingStock 최적화
- CartSidebart.tsx → CartSidebar.tsx로 파일명 오타 수정
- CartSidebar 컴포넌트 내 any 타입 제거 및 명확한 타입 정의
- getRemainingStock 연산을 useMemo로 감싸 계산 최적화
