+++ src/basic/components/customer/cart/CartSidebar.tsx
@@ -1,13 +1,13 @@
-import { CartItem } from "../../../../types";
+import { CartItem, Coupon } from "../../../../types";
 import CartItemBox from "./CartItemBox";
 import { CouponSection } from "../CouponSection";
 import { PaymentSection } from "../PaymentSection";
-import { BagIcon } from "../../elements/Icons";
+import { BagIcon, EmptyCartIcon } from "../../elements/Icons";
 
 interface CartSidebarProps {
   cart: CartItem[];
-  coupons: any[];
-  selectedCoupon: any;
+  coupons: Coupon[];
+  selectedCoupon: Coupon | null;
   totals: { totalBeforeDiscount: number; totalAfterDiscount: number };
   calculateItemTotal: (item: CartItem) => number;
   removeFromCart: (productId: string) => void;
@@ -28,9 +28,7 @@ export const CartSidebar = ({ cart, coupons, selectedCoupon, totals, calculateIt
 
         {cart.length === 0 ? (
           <div className="text-center py-8">
-            <svg className="w-16 h-16 text-gray-300 mx-auto mb-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1} d="M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z" />
-            </svg>
+            <EmptyCartIcon />
             <p className="text-gray-500 text-sm">장바구니가 비어있습니다</p>
           </div>
         ) : (
